chore: landing fixes for #8794 (object/mod.rs split, raw-handle ceilings)
