+++ scripts/raw_handle_debt_files.txt
@@ -94,7 +94,7 @@
 12 crates/perry-runtime/src/object/descriptors.rs
 5 crates/perry-runtime/src/object/field_get_set/enumeration.rs
 2 crates/perry-runtime/src/object/field_get_set/get_field_by_name.rs
-3 crates/perry-runtime/src/object/field_set_by_name/fast_paths.rs
+8 crates/perry-runtime/src/object/field_set_by_name/fast_paths.rs
 11 crates/perry-runtime/src/object/field_set_by_name/tail.rs
 2 crates/perry-runtime/src/object/global_this/ctor_thunks.rs
 9 crates/perry-runtime/src/object/global_this/populate.rs
@@ -155,3 +155,4 @@
 19 crates/perry-runtime/src/wasi.rs
 6 crates/perry-runtime/src/weakref.rs
 20 crates/perry-runtime/src/webassembly.rs
+6 crates/perry-runtime/src/proxy/put_value.rs
